requirements.txt
# Core dependencies
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0

# Optional: For local models
# transformers>=4.30.0
# torch>=2.0.0

